api-location-livre/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
email-validator==2.2.0
httpx==0.27.2
pytest==8.3.3
pytest-cov==5.0.0
psycopg[binary]
black==24.10.0
ruff==0.6.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

bcrypt==4.0.1

python-multipart==0.0.9
requests==2.32.3

